+++ codex-rs/exec/src/event_processor_with_human_output.rs
@@ -1,6 +1,6 @@
 use codex_common::elapsed::format_elapsed;
 use codex_core::config::Config;
-use codex_core::handle_update_plan::UpdatePlanArgs;
+use codex_core::plan_tool::UpdatePlanArgs;
 use codex_core::protocol::AgentMessageDeltaEvent;
 use codex_core::protocol::AgentMessageEvent;
 use codex_core::protocol::AgentReasoningDeltaEvent;
